Move ReportLevel enum to datatypes.h
To avoid circular dependencies
